Update timelineView
All click actions are exclusive, directly returning in the loop
Removed graphicsitem  members to only keep one and a drag mode.

Timeline connects the view to managers directly

+++ Timeline/ABManager.cpp
@@ -11,7 +11,7 @@ ABManager::ABManager(QGraphicsScene *scene, TimelineMath *mathManager, QObject *
 {
 }
 
-void ABManager::dragMarker(ABMarkerItem* marker, const int pos)
+void ABManager::dragMarker(ABMarkerItem* marker, const double pos)
 {
     changeMarkerTime(marker, p_mathManager->posToTimeSnapped(pos));
 }
